[refactor] make Accred independent from Accreditation for editing
This makes the accreds_controller much more standard because all the
fields that are needed for editing are cached in the Accred objects.
This is also yet another attempt of fixing the name clash issue.
I should probably converto all the "position" columns used for sorting
into "order" so we no longer clash with "position" as in "job position".
For the moment I will use "role" as attribute for the Accred class to
store the job position.