[feat] wsgetpeople POC scipers,groups,units
This starts to look like the output of legacy counterpart.
We are still only checking that the units are all at the same level but
we are not limiting the choice to leaf units.
Regarding groups, we need the group id (Sxxxxx) for calling api group members but
api is too verbose and cannot be forced to return only by exact name match.
Therefore, for the moment I rely on LDAP for converting group name into group id.