[feat] add a short expertise field to profile
The idea is to use it for just few keywords regarding person's expertise but still keep the
previous expertise box. Many people have very long expertise records in legacy people and
these do not fit the idea of short keywords-like field. We started by moving the box into
profile (see pr #46) but then we realised that this were not feasible with our current data set.