[fix][chore] workaround for keycloak
in dev it is still useful to have keycloak in order to be able to impersonate Natalie.
But, while entraid returns the sciper as uniqueid, keycloak needs some extra
configuration which I don't have the patience to figure out. Therefore I add
a workaround to determine the sciper number from the email address which is
instead present also in the keycloak token.