[chore] prepare for eventually using Lindo's opdo spyware
It remains to understand why the version that takes the Gem directly from git
does not work. I suspect it is saved in the wrong path.
The way it can work currently is to clone the gem in vendor/gems (to be automated
in Makefile if we decide to adopt it).