[refact] generic audience_limitable + !popup visibility
This commit introduces the possibility to attach visibility not only to models, but also to single properties of a model provided there is a PROPERTYNAME_visibility column in the DB.
This requires explicitly calling the concern setup because i might be needed more than once for a given model.

I really do not like the popup for selecting visibility. I also think it is not compatible with touch devices. Therefore I go back to my previous design where the buttons for switching visibility are always visible.