[ui] moving all edits to a custom modal editor
Initially intended for boxes, it actually started with simple models.
Now boxes are slightly more complicated but I think it is a nice
way to go.
In particular, this will enable a sort of wysiwyg editing mode
where the sorting is done on a TOC column and the content looks
almost like the public page but with edit/delete buttons on hover.