[refact] yet another version of the name change UI. Fixes #188
NM spoke with HR and they confirm that the current rules are too complicated. EGA agreed to be the
unique bridge for all usual name changes. Therefore now we do not need to change the messages based
on the type of person.
We decided to always display both the official and the usual name and have dedicated modal forms
instead of a single modal with tabs. In the usual name modal there will be two parts: one that
is always shown that concerns special name changes that will send the request to EGA, and the
other for the standard self-service name changes.