[fix] #182 too many accreditors in function change form
This adds yet another horror by memoizing the full accreditation within the accred. Yes, I know,
it is a bit convoluted: accreditation contains accred which contains accreditation...
One day I'd like to find a nice way to get rid of this horrible accreditation/accred duality.

It also fix the fact that students have no accreditors because (I presume) they are under
the SAC responsibility. Therefore, they should not be given the option to request for a
position change. In fact, how else should be called a student ?