[feat] nginx server proxying legacy app directly from the servers
This involves a big refactoring of the deployment code for openshit.
We now have an unified system of tags for uniquely addressing
the various components of the deployment. This allows for an easier
deployment of different versions and or run environments and, at
this stage, have people-next cohabit with people.