[WIP] putting this in standby for now. Test needed
Once modified 'M' or 'F' into 'X' there is no way back. Therefore, for now
we still need to use the legacy server for changing inclusivity.
Asked IAM to provide us with official_gender or a property that we can toggle.
This should work using the new api for names and the old server for inclusivity
but I don't know how to test. Let's let it mature for a while and then we'll see.