[refact] phd importer and related stuff
The previous approach of distinguishing past from current PhDs base on
the presence of the `date` field was not very reliable.
Therefore, we add a dedicated field in the PhD model and fill it based
on the request sent to Oasis: either /student or /alumni.
This also fix a recent bug introduced with BulkPeople which returned
a very incomplete list of person when looking for old PhDs because
they are no longer accredited.