[chore] move jobs worker to next release
As long as next release is backward compatible with prod release, this
enable testing new code that requires updated jobs which are not
yet present in the prod version.
We should definitely keep next for very small changes and rely
on test for deeper ones. Therefore this is a temporary fix while
waiting for the test env to be deployed.