[feat] default selectable property
It have been decided to relax the fact of forcing the presence of origin and category for Award.
The first option where to extend WithSelectableProperties in order to add an "optional" flag.
Instead I opted for the easiest option of adding a default property (that can be created with _empty_
value if we really want to give the possibility of having blank properties.